Works with real filter lifecycle
The filter lifecycle defined in the Javadoc for Filter class is a little
misleading.  This new implementation respects the real filter lifecycle
and tests to make sure that it does.

Also adds a small amount of log output to help in debugging, though this
will likely go away very, very soon.