Adds a real implementation of a dynamically-loaded filter
A DeployedFilter can be sent to the HBase region servers via HDFS and
loaded dynamically via a class loader that the DeployedFilterManager
manages.  Client code dictates which jar a filter lives in, and thus
which version of the filter gets deployed to the cluster.

When `mvn package` is run, an additional jar will be created that
contains the DeployedFilter and necessary classes.  This jar (named
deployed-filter-${project.version}-bin.jar) can be deployed to the HBase
region servers and used to dynamically load other filters.