Adds metrics collection for the depoyed filters
The metrics include the numer of filter classes and corresponding class
loaders that have been instantiated as well as the number of filter
classes that have been dynamically loaded.  Additionally, the garbage
collection of the deployed filters and their class loaders is tracked
so that memory leaks can be detected.