Adds new Selector interface to augment Filter
A Selector inverses the contract of Filter (and simplifies it a bit) to
allow writing of HBase filter functionality that is easier to
understand.

The AbstractStatelessSelector implements the Writable methods so that
you do not have to, if you have no serialized state in your Selector.

Adds a Selector-based implementation of the RowKeyInSetFilter called
RowKeyInSetSelector.